chore(jsnam) : file refactor
