+++ src/app/providers/RouterProvider.tsx
@@ -5,10 +5,8 @@ import Community from 'pages/community';
 import Create from 'pages/create/';
 import Demo from 'pages/demo';
 import MusicDetail from 'pages/detail';
-import Home from 'pages/home/Home';
-import MemberCommunity from 'pages/memberCommunity';
-import MemberInfo from 'pages/memberInfo';
-import MemberReview from 'pages/memberReview';
+import Home from 'pages/home';
+import { MemberCommunity, MemberInfo, MemberReview } from 'pages/memberInfo';
 import NotFound from 'pages/notFound/NotFound';
 import PlayList from 'pages/playList';
 
@@ -24,9 +22,9 @@ export const AppRouter = () => {
       <Route path={ROUTES.CREATE} element={<Create />} />
       <Route path={ROUTES.PLAYLIST} element={<PlayList />} />
 
-      <Route path={ROUTES.MEMBERINFO} element={<MemberInfo />} />
-      <Route path={ROUTES.MEMBERINFOCOMMUNITY} element={<MemberCommunity />} />
-      <Route path={ROUTES.MEMBERINFOREVIEW} element={<MemberReview />} />
+      <Route path={ROUTES.MEMBERINFO.MEMBERINFO} element={<MemberInfo />} />
+      <Route path={ROUTES.MEMBERINFO.MEMBERINFOCOMMUNITY} element={<MemberCommunity />} />
+      <Route path={ROUTES.MEMBERINFO.MEMBERINFOREVIEW} element={<MemberReview />} />
 
       <Route path={ROUTES.ADMIN.NOTICE} element={<AdminNoticePage />} />
       <Route path={ROUTES.ADMIN.USER} element={<AdminUserPage />} />
