--- src/pages/community/Community.tsx
@@ -1,12 +1,0 @@
-import { Link } from 'react-router-dom';
-
-function Community() {
-  return (
-    <div>
-      <h1>커뮤니티 페이지</h1>
-      <Link to="/">홈페이지로 이동</Link>
-    </div>
-  );
-}
-
-export default Community;
