+++ src/pages/community/ui/RecommendedPost.tsx
