+++ src/pages/community/ui/index.tsx
@@ -1,12 +1,13 @@
 import styled from '@emotion/styled';
 
-import { COMMUNITY_ITEM } from './constants';
+import { COMMUNITY_ITEM } from './model';
 import RecommendedPost from './RecommendedPost';
 import RecommendedPostList from './RecommendedPostList';
 
 export const Community = () => {
   const MainContents = styled.div`
     width: 100%;
+    height: 2832px;
   `;
 
   return (
