+++ src/pages/home/ui/types/index.ts
