+++ src/pages/memberInfo/member/index.tsx
@@ -1,0 +1,11 @@
+import { MainLayout } from 'widgets/ui/Layout';
+
+import { MemberInfoPage } from './ui';
+
+export const MemberInfo = () => {
+  return (
+    <MainLayout>
+      <MemberInfoPage></MemberInfoPage>
+    </MainLayout>
+  );
+};
