+++ src/pages/memberInfo/member/ui/MemberCommunity.tsx
@@ -2,19 +2,19 @@ import styled from '@emotion/styled';
 import moment from 'moment';
 moment.locale('ko');
 
+import { CommunityMusicInfo } from 'pages/home/ui/types';
+
 import { commonStyles } from 'shared/styles/common';
-import { StarRatingInput } from 'shared/ui/Input/StarRatingInput';
-import { CommunityMusicInfo } from 'shared/ui/Main/types';
 
 interface CommunityMusicProps {
   CommunityMusicInfo: CommunityMusicInfo;
 }
 
-export const MemberReviewList = ({ CommunityMusicInfo }: CommunityMusicProps) => {
+export const MemberCommunity = ({ CommunityMusicInfo }: CommunityMusicProps) => {
   return (
     <MemberContainer>
       <TitleBlock>
-        <PageTitle>나의 별점 및 리뷰</PageTitle>
+        <PageTitle>나의 음악 추천 게시글</PageTitle>
         <MoreTitle>더보기</MoreTitle>
       </TitleBlock>
 
@@ -25,19 +25,21 @@ export const MemberReviewList = ({ CommunityMusicInfo }: CommunityMusicProps) =>
               <div key={item.id}>
                 <CommunityListWrapper>
                   <CommunityList>
-                    <StarRatingWrapper>
-                      <StarRatingInput value={5} enabled={false} />
-                    </StarRatingWrapper>
+                    <ListImg src={item.img} alt={item.title} />
                     <ListContent>
-                      <ListImg src={item.img} alt={item.title} />
                       <ContentInfo>
                         <ContentsSongName>{item.title}</ContentsSongName>
                         <ContentsSongDescription>{item.description}</ContentsSongDescription>
                       </ContentInfo>
+
+                      <ContentTitleBlock>
+                        <ContentTitle>게시글 제목</ContentTitle>
+                      </ContentTitleBlock>
+
+                      <ActivityInfo>
+                        <ActivityStatus>{moment(item.date).format('YYYY-MM-DD')}</ActivityStatus>
+                      </ActivityInfo>
                     </ListContent>
-                    <ActivityInfo>
-                      <ActivityStatus>{moment(item.date).format('YYYY-MM-DD')}</ActivityStatus>
-                    </ActivityInfo>
                   </CommunityList>
                 </CommunityListWrapper>
               </div>
@@ -106,6 +108,7 @@ const CommunityListWrapper = styled.div`
 `;
 
 const CommunityList = styled.div`
+  padding: 24px;
   width: 1160px;
   height: 112px;
   display: flex;
@@ -113,12 +116,6 @@ const CommunityList = styled.div`
   align-items: center;
   border: 1px solid ${({ theme }) => theme.colors[400]};
   border-radius: 1px;
-  padding: 24px;
-`;
-
-const StarRatingWrapper = styled.div`
-  width: 148px;
-  height: 28px;
 `;
 
 const ListImg = styled.img`
@@ -128,15 +125,15 @@ const ListImg = styled.img`
 `;
 
 const ListContent = styled.div`
-  width: 968px;
+  width: 1040px;
   height: 64px;
   display: flex;
-  /* justify-content: center; */
+  justify-content: center;
   align-items: center;
 `;
 
 const ContentInfo = styled.div`
-  width: 234px;
+  width: 160px;
   height: 64px;
   display: flex;
   flex-direction: column;
@@ -156,12 +153,28 @@ const ContentsSongDescription = styled.div`
   ${commonStyles.limitText};
 `;
 
+const ContentTitleBlock = styled.div`
+  width: 708px;
+  height: 28px;
+  margin-left: 16px;
+`;
+
+const ContentTitle = styled.div`
+  ${({ theme }) => theme.fonts.wantedSans.B3};
+  color: ${({ theme }) => theme.colors[100]};
+  ${commonStyles.limitText};
+
+  &::before {
+    content: '·';
+    margin-right: 4px;
+  }
+`;
+
 const ActivityInfo = styled.div`
   display: flex;
   flex-direction: column;
   justify-content: center;
-  gap: 8px;
-  width: 120px;
+  width: 200px;
   height: 52px;
   margin-left: 16px;
 `;
@@ -169,9 +182,5 @@ const ActivityInfo = styled.div`
 const ActivityStatus = styled.div`
   ${({ theme }) => theme.fonts.wantedSans.B6};
   color: ${({ theme }) => theme.colors[200]};
-  max-width: 192px;
-  overflow: hidden;
-  text-overflow: ellipsis;
-  white-space: nowrap;
   text-align: right;
 `;
