+++ src/shared/ui/index.ts
@@ -2,14 +2,9 @@ export { PlayListPage } from '../../widgets/ui/PlayList';
 export * from './ArrowButton';
 export * from './Button';
 export { Chip } from './Chip';
-export { Community } from './Community';
 export * from './Input';
-export { Main } from './Main';
 export { Modal, OuterCloseModal } from './Modal';
 export { MoreButton } from './MoreButton';
 export { Pagination } from './Pagination';
 export * from './Table';
 export { YoutubeIframe } from './YoutubeIframe';
-export { MemberInfoPage } from './MemberInfo';
-export { MemberInfoCommunity } from './MemberCommunity';
-export { MemberInfoReview } from './MemberReview';
